feat: breaks up some of the logic into their respective classes, adds typing, adjusted for db calls and loop event, some generic stuff
