+++ config/constants.py
@@ -1,0 +1,4 @@
+TLM_API_URL = "https://api.treasureland.market"
+TLM_API_VERSION = "v2/v1"
+BSC_CHAIN_ID = "56"
+BUNNY_CONTRACT = "0xdf7952b35f24acf7fc0487d01c8d5690a60dba07"
